requirements.txt
# Core RL libraries
gymnasium>=0.29.0
stable-baselines3>=2.0.0
ray[rllib]>=2.8.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Visualization and UI
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
tensorboard>=2.13.0
wandb>=0.15.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
